Integer to BigInt in db schema i32 to i64 in rust related object

+++ src/vit-servicing-station/migrations/2020-05-22-112032_initial/up.sql
@@ -16,9 +16,9 @@ CREATE TABLE proposals (
     proposer_url VARCHAR NOT NULL,
     chain_proposal_id VARCHAR NOT NULL,
     chain_voteplan_id VARCHAR NOT NULL,
-    chain_proposal_index INTEGER NOT NULL,
-    chain_vote_start_time INTEGER NOT NULL,
-    chain_vote_end_time INTEGER NOT NULL,
-    chain_committee_end_time INTEGER NOT NULL,
+    chain_proposal_index BIGINT NOT NULL,
+    chain_vote_start_time BIGINT NOT NULL,
+    chain_vote_end_time BIGINT NOT NULL,
+    chain_committee_end_time BIGINT NOT NULL,
     chain_vote_options VARCHAR NOT NULL
 )
